requirements.txt
Flask
Flask-SQLAlchemy
PyMySQL
gunicorn
werkzeug
Flask-Migrate
cryptography